requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
sqlglot==25.20.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==2.0.1
faiss-cpu==1.8.0.post1
jinja2==3.1.4
rich==13.7.1
pytest==8.3.2
httpx==0.27.0